+++ branches/5.2.0/src/psimage.c
@@ -418,7 +418,7 @@ int GMT_psimage (void *V_API, int mode, void *args)
 			I->header->n_bands = 3;
 		}
 		else if (I->header->n_bands == 4) { /* RGBA image, with a color map */
-			if (!Ctrl->G.active) has_trans = find_unique_color (GMT, I->data, I->header->nm, &r, &g, &b);
+			if (!Ctrl->G.active) has_trans = find_unique_color (GMT, I->data, (int)I->header->nm, &r, &g, &b);
 			for (j = n = 0; j < 4 * (int)I->header->nm; j++) { /* Reduce image from 32- to 24-bit */
 				if (has_trans && I->data[j+3] == 0)
 					I->data[n++] = r, I->data[n++] = g, I->data[n++] = b, j+=3;
